+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Tidarr notable changes.
 - [Lidarr] Apply PUID and PGID to Lidarr downloaded files
 - [Front] Add single download button in no-download mode #684
 - [Front] Add toggle finished items button in queue list
+- [Docker] Load custom `/config/requirements.txt` file to extend your scripts
 
 ## 📦 1.2.0
 ### 🐛 Fixed